Move `cache.ts` to `features/` directory.
