+++ src/main.ts
@@ -3,7 +3,7 @@ import * as core from '@actions/core'
 import * as graalvm from './graalvm'
 import {isFeatureAvailable as isCacheAvailable} from '@actions/cache'
 import {join} from 'path'
-import {restore} from './cache'
+import {restore} from './features/cache'
 import {setUpDependencies} from './dependencies'
 import {setUpGUComponents} from './gu'
 import {setUpMandrel} from './mandrel'
